appsocketapi/requirements.txt
fastapi>=0.111.0,<0.113.0
pydantic>=2.7.0,<3.0.0
dapr
websockets
azure-identity